+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-03-02
+### Added
+- `CI` **Lint** workload.
+- New flag `--undef` to set a value for undefined location.
+
+### Fixed
+- All lint errors.
+
+### Changed
+- Keep checking location every round despite interfaces number change if the last check failed.
+
 ## [0.4.0] - 2025-02-27
 ### Added
 - `CHANGELOG.md`.